requirements.txt
acres==0.3.0
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dotenv==0.9.9
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydot==3.0.4
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
